+++ WorldOfImagination/Maker.Rise/Utils/Colision.cs
@@ -2,7 +2,7 @@
 {
     public static class Colision
     {
-        public static bool Check(int x1, int y1, int w1, int h1, int x2, int y2, int w2, int h2)
+        public static bool Check(float x1, float y1, float w1, float h1, float x2, float y2, float w2, float h2)
         {
             return x1 < x2 + w2 &&
                    x2 < x1 + w1 &&
